agent_v2/requirements.txt
langgraph
langchain
langchain-groq
python-dotenv
faiss-cpu
langchain-community
